requirements.txt
absl-py==0.3.0
grpcio==1.15.0
numpy==1.13.1
protobuf==3.6.1
psutil==5.4.0
numpy==1.13.1
